requirements.txt
reportlab
babel
boto3
dateutils
requests
aws-lambda-logging
